refactor: Испралено название файла хука
