+++ src/pages/forgot-password-page/main/forgot-password-form/forgot-password-form.tsx
@@ -8,7 +8,7 @@ import styles from './forgot-password-form.module.css';
 
 import ButtonLoader from '../../../../components/button-loader/button-loader';
 import { useAppDispatch, useAppSelector } from '../../../../hooks/rtk-hooks';
-import useFormAndValidation from '../../../../hooks/use-form-and-validation';
+import useFormAndValidation from '../../../../hooks/useFormAndValidation';
 import { forgotPassword } from '../../../../services/reducers/forgot-password';
 import { openModal } from '../../../../services/reducers/modal';
 import { selectForgotPasswordRequest } from '../../../../services/selectors';
