+++ src/pages/login-page/main/login-form/login-form.tsx
@@ -7,7 +7,7 @@ import styles from './login-form.module.css';
 
 import ButtonLoader from '../../../../components/button-loader/button-loader';
 import { useAppDispatch, useAppSelector } from '../../../../hooks/rtk-hooks';
-import useFormAndValidation from '../../../../hooks/use-form-and-validation';
+import useFormAndValidation from '../../../../hooks/useFormAndValidation';
 import { openModal } from '../../../../services/reducers/modal';
 import { login } from '../../../../services/reducers/profile';
 import { selectUserDataRequest } from '../../../../services/selectors';
